fix: prevent integer overflow in reflect coordinate calculation
Use long arithmetic (2L * (size - 1)) instead of int multiplication
to prevent potential overflow when size is large. Applied to:
- Scale.cs
- Rotation.cs
- Affine.cs

+++ src/Augmentation/Image/Geometric/Scale.cs
@@ -198,7 +198,8 @@ public class Scale<T> : SpatialAugmentationBase<T, ImageTensor<T>>
 
         // Use modular arithmetic for deterministic reflection
         // Period is 2 * (size - 1) for reflection pattern
-        int period = 2 * (size - 1);
+        // Use long arithmetic to prevent overflow for large sizes
+        long period = 2L * (size - 1);
 
         // Handle negative coordinates
         coord = Math.Abs(coord);
